+++ vkcs/resource_vkcs_db_cluster.go
@@ -2,8 +2,10 @@ package vkcs
 
 import (
 	"context"
+	"errors"
 	"fmt"
 	"log"
+	"strings"
 
 	"github.com/hashicorp/go-cty/cty"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
@@ -656,278 +658,85 @@ func resourceDatabaseClusterRead(ctx context.Context, d *schema.ResourceData, me
 
 func resourceDatabaseClusterUpdate(ctx context.Context, d *schema.ResourceData, meta interface{}) diag.Diagnostics {
 	config := meta.(configer)
-	DatabaseV1Client, err := config.DatabaseV1Client(getRegion(d, config))
+	dbClient, err := config.DatabaseV1Client(getRegion(d, config))
 	if err != nil {
 		return diag.Errorf("Error creating VKCS database client: %s", err)
 	}
 
+	clusterID := d.Id()
 	stateConf := &resource.StateChangeConf{
 		Pending:    []string{string(dbClusterStatusBuild)},
 		Target:     []string{string(dbClusterStatusActive)},
-		Refresh:    databaseClusterStateRefreshFunc(DatabaseV1Client, d.Id(), nil),
+		Refresh:    databaseClusterStateRefreshFunc(dbClient, d.Id(), nil),
 		Timeout:    d.Timeout(schema.TimeoutCreate),
 		Delay:      dbInstanceDelay,
 		MinTimeout: dbInstanceMinTimeout,
 	}
+	updateCtx := &dbResourceUpdateContext{
+		Ctx:       ctx,
+		Client:    dbClient,
+		D:         d,
+		StateConf: stateConf,
+	}
 
 	if d.HasChange("configuration_id") {
-		old, new := d.GetChange("configuration_id")
-
-		var detachConfigurationOpts dbClusterDetachConfigurationGroupOpts
-		detachConfigurationOpts.ConfigurationDetach.ConfigurationID = old.(string)
-		err := dbClusterAction(DatabaseV1Client, d.Id(), &detachConfigurationOpts).ExtractErr()
+		err = databaseClusterActionUpdateConfiguration(updateCtx)
 		if err != nil {
-			return diag.FromErr(err)
-		}
-		log.Printf("Detaching configuration %s from vkcs_db_cluster %s", old, d.Id())
-
-		_, err = stateConf.WaitForStateContext(ctx)
-		if err != nil {
-			return diag.Errorf("error waiting for vkcs_db_cluster %s to become ready: %s", d.Id(), err)
-		}
-
-		if new != "" {
-			var attachConfigurationOpts dbClusterAttachConfigurationGroupOpts
-			attachConfigurationOpts.ConfigurationAttach.ConfigurationID = new.(string)
-			err := dbClusterAction(DatabaseV1Client, d.Id(), &attachConfigurationOpts).ExtractErr()
-			if err != nil {
-				return diag.FromErr(err)
-			}
-			log.Printf("Attaching configuration %s to vkcs_db_cluster %s", new, d.Id())
-
-			_, err = stateConf.WaitForStateContext(ctx)
-			if err != nil {
-				return diag.Errorf("error waiting for vkcs_db_cluster %s to become ready: %s", d.Id(), err)
-			}
+			return databaseClusterUpdateProcessError(err, clusterID)
 		}
 	}
 
 	if d.HasChange("volume_size") {
-		_, new := d.GetChange("volume_size")
-		var resizeVolumeOpts dbClusterResizeVolumeOpts
-		resizeVolumeOpts.Resize.Volume.Size = new.(int)
-		err := dbClusterAction(DatabaseV1Client, d.Id(), &resizeVolumeOpts).ExtractErr()
+		err = databaseClusterActionResizeVolume(updateCtx, "")
 		if err != nil {
-			return diag.FromErr(err)
-		}
-		log.Printf("Resizing volume from vkcs_db_cluster %s", d.Id())
-
-		stateConf.Pending = []string{string(dbClusterStatusResize)}
-		stateConf.Target = []string{string(dbClusterStatusActive)}
-
-		_, err = stateConf.WaitForStateContext(ctx)
-		if err != nil {
-			return diag.Errorf("error waiting for vkcs_db_cluster %s to become ready: %s", d.Id(), err)
+			return databaseClusterUpdateProcessError(err, clusterID)
 		}
 	}
 
 	if d.HasChange("flavor_id") {
-		var resizeOpts dbClusterResizeOpts
-		resizeOpts.Resize.FlavorRef = d.Get("flavor_id").(string)
-		err := dbClusterAction(DatabaseV1Client, d.Id(), &resizeOpts).ExtractErr()
+		err = databaseClusterActionResizeFlavor(updateCtx, "")
 		if err != nil {
-			return diag.FromErr(err)
-		}
-		log.Printf("Resizing flavor from vkcs_db_cluster %s", d.Id())
-
-		stateConf.Pending = []string{string(dbClusterStatusResize)}
-		stateConf.Target = []string{string(dbClusterStatusActive)}
-
-		_, err = stateConf.WaitForStateContext(ctx)
-		if err != nil {
-			return diag.Errorf("error waiting for vkcs_db_cluster %s to become ready: %s", d.Id(), err)
+			return databaseClusterUpdateProcessError(err, clusterID)
 		}
 	}
 
 	if d.HasChange("disk_autoexpand") {
-		_, new := d.GetChange("disk_autoexpand")
-		autoExpandProperties, err := extractDatabaseAutoExpand(new.([]interface{}))
+		err = databaseClusterUpdateDiskAutoexpand(updateCtx)
 		if err != nil {
-			return diag.Errorf("unable to determine vkcs_db_cluster disk_autoexpand")
-		}
-		var autoExpandOpts dbClusterUpdateAutoExpandOpts
-		if autoExpandProperties.AutoExpand {
-			autoExpandOpts.Cluster.VolumeAutoresizeEnabled = 1
-		} else {
-			autoExpandOpts.Cluster.VolumeAutoresizeEnabled = 0
-		}
-		autoExpandOpts.Cluster.VolumeAutoresizeMaxSize = autoExpandProperties.MaxDiskSize
-		err = dbClusterUpdateAutoExpand(DatabaseV1Client, d.Id(), &autoExpandOpts).ExtractErr()
-		if err != nil {
-			return diag.FromErr(err)
-		}
-
-		stateConf.Pending = []string{string(dbClusterStatusUpdating)}
-		stateConf.Target = []string{string(dbClusterStatusActive)}
-
-		_, err = stateConf.WaitForStateContext(ctx)
-		if err != nil {
-			return diag.Errorf("error waiting for vkcs_db_cluster %s to become ready: %s", d.Id(), err)
+			return databaseClusterUpdateProcessError(err, clusterID)
 		}
 	}
 
 	if d.HasChange("wal_volume") {
-		old, new := d.GetChange("wal_volume")
-		walVolumeOptsNew, err := extractDatabaseWalVolume(new.([]interface{}))
+		err = databaseClusterActionResizeWalVolume(updateCtx, "")
 		if err != nil {
-			return diag.Errorf("unable to determine vkcs_db_cluster wal_volume")
+			return databaseClusterUpdateProcessError(err, clusterID)
 		}
-
-		walVolumeOptsOld, err := extractDatabaseWalVolume(old.([]interface{}))
-		if err != nil {
-			return diag.Errorf("unable to determine vkcs_db_cluster wal_volume")
-		}
-
-		if walVolumeOptsNew.Size != walVolumeOptsOld.Size {
-			var resizeWalVolumeOpts dbClusterResizeWalVolumeOpts
-			resizeWalVolumeOpts.Resize.Volume.Size = walVolumeOptsNew.Size
-			resizeWalVolumeOpts.Resize.Volume.Kind = "wal"
-			err = dbClusterAction(DatabaseV1Client, d.Id(), &resizeWalVolumeOpts).ExtractErr()
-			if err != nil {
-				return diag.FromErr(err)
-			}
-
-			stateConf.Pending = []string{string(dbClusterStatusResize)}
-			stateConf.Target = []string{string(dbClusterStatusActive)}
-
-			_, err = stateConf.WaitForStateContext(ctx)
-			if err != nil {
-				return diag.Errorf("error waiting for vkcs_db_cluster %s to become ready: %s", d.Id(), err)
-			}
-		}
-
 	}
+
 	if d.HasChange("wal_disk_autoexpand") {
-		_, new := d.GetChange("wal_disk_autoexpand")
-		walAutoExpandProperties, err := extractDatabaseAutoExpand(new.([]interface{}))
+		err = databaseClusterUpdateWalDiskAutoexpand(updateCtx)
 		if err != nil {
-			return diag.Errorf("unable to determine vkcs_db_cluster wal_disk_autoexpand")
-		}
-		var walAutoExpandOpts dbClusterUpdateAutoExpandWalOpts
-		if walAutoExpandProperties.AutoExpand {
-			walAutoExpandOpts.Cluster.WalVolume.VolumeAutoresizeEnabled = 1
-		} else {
-			walAutoExpandOpts.Cluster.WalVolume.VolumeAutoresizeEnabled = 0
-		}
-		walAutoExpandOpts.Cluster.WalVolume.VolumeAutoresizeMaxSize = walAutoExpandProperties.MaxDiskSize
-		err = dbClusterUpdateAutoExpand(DatabaseV1Client, d.Id(), &walAutoExpandOpts).ExtractErr()
-		if err != nil {
-			return diag.FromErr(err)
-		}
-
-		stateConf.Pending = []string{string(dbClusterStatusUpdating)}
-		stateConf.Target = []string{string(dbClusterStatusActive)}
-
-		_, err = stateConf.WaitForStateContext(ctx)
-		if err != nil {
-			return diag.Errorf("error waiting for vkcs_db_cluster %s to become ready: %s", d.Id(), err)
+			return databaseClusterUpdateProcessError(err, clusterID)
 		}
 	}
 
 	if d.HasChange("capabilities") {
-		_, newCapabilities := d.GetChange("capabilities")
-		newCapabilitiesOpts, err := extractDatabaseCapabilities(newCapabilities.([]interface{}))
+		err = databaseClusterActionApplyCapabilities(updateCtx)
 		if err != nil {
-			return diag.Errorf("unable to determine vkcs_db_cluster capability")
-		}
-		var applyCapabilityOpts dbClusterApplyCapabilityOpts
-		applyCapabilityOpts.ApplyCapability.Capabilities = newCapabilitiesOpts
-
-		err = dbClusterAction(DatabaseV1Client, d.Id(), &applyCapabilityOpts).ExtractErr()
-
-		if err != nil {
-			return diag.Errorf("error applying capability to vkcs_db_cluster %s: %s", d.Id(), err)
-		}
-
-		applyCapabilityClusterConf := &resource.StateChangeConf{
-			Pending:    []string{string(dbClusterStatusCapabilityApplying), string(dbClusterStatusBuild)},
-			Target:     []string{string(dbClusterStatusActive)},
-			Refresh:    databaseClusterStateRefreshFunc(DatabaseV1Client, d.Id(), &newCapabilitiesOpts),
-			Timeout:    d.Timeout(schema.TimeoutCreate),
-			Delay:      dbInstanceDelay,
-			MinTimeout: dbInstanceMinTimeout,
-		}
-		log.Printf("[DEBUG] Waiting for cluster to become ready after applying capability")
-		_, err = applyCapabilityClusterConf.WaitForStateContext(ctx)
-		if err != nil {
-			return diag.Errorf("error applying capability to vkcs_db_cluster %s: %s", d.Id(), err)
+			return databaseClusterUpdateProcessError(err, clusterID)
 		}
 	}
 
 	if d.HasChange("cluster_size") {
 		old, new := d.GetChange("cluster_size")
-		if new.(int) > old.(int) {
-			opts := make([]dbClusterGrowOpts, new.(int)-old.(int))
-
-			volumeSize := d.Get("volume_size").(int)
-			growOpts := dbClusterGrowOpts{
-				Keypair:          d.Get("keypair").(string),
-				AvailabilityZone: d.Get("availability_zone").(string),
-				FlavorRef:        d.Get("flavor_id").(string),
-				Volume:           &volume{Size: &volumeSize, VolumeType: d.Get("volume_type").(string)},
-			}
-			if v, ok := d.GetOk("wal_volume"); ok {
-				walVolumeOpts, err := extractDatabaseWalVolume(v.([]interface{}))
-				if err != nil {
-					return diag.Errorf("unable to determine vkcs_db_cluster wal_volume")
-				}
-				growOpts.Walvolume = &walVolume{
-					Size:       &walVolumeOpts.Size,
-					VolumeType: walVolumeOpts.VolumeType,
-				}
-			}
-			for i := 0; i < len(opts); i++ {
-				opts[i] = growOpts
-			}
-			growClusterOpts := dbClusterGrowClusterOpts{
-				Grow: opts,
-			}
-			err = dbClusterAction(DatabaseV1Client, d.Id(), &growClusterOpts).ExtractErr()
-
-			if err != nil {
-				return diag.Errorf("error growing vkcs_db_cluster %s: %s", d.Id(), err)
-			}
-			stateConf.Pending = []string{string(dbClusterStatusGrow)}
-			stateConf.Target = []string{string(dbClusterStatusActive)}
-
-			_, err = stateConf.WaitForStateContext(ctx)
-			if err != nil {
-				return diag.Errorf("error waiting for vkcs_db_cluster %s to become ready: %s", d.Id(), err)
-			}
-		} else {
-			rawShrinkOptions := d.Get("shrink_options").([]interface{})
-			shrinkOptions := expandDatabaseClusterShrinkOptions(rawShrinkOptions)
-			if len(shrinkOptions) > 0 && len(shrinkOptions) != new.(int) {
-				return diag.Errorf("invalid shrink options: number of instances in shrink options should equal new cluster size")
-			}
-
-			cluster, err := dbClusterGet(DatabaseV1Client, d.Id()).extract()
-			if err != nil {
-				return diag.FromErr(checkDeleted(d, err, "Error retrieving vkcs_db_cluster"))
-			}
-
-			toDelete := old.(int) - new.(int)
-			ids, err := databaseClusterDetermineShrinkedInstances(toDelete, shrinkOptions, cluster.Instances)
-			if err != nil {
-				return diag.Errorf("error determining instances to shrink: %s", err)
-			}
-
-			shrinkClusterOpts := dbClusterShrinkClusterOpts{
-				Shrink: ids,
-			}
-
-			err = dbClusterAction(DatabaseV1Client, d.Id(), &shrinkClusterOpts).ExtractErr()
-
-			if err != nil {
-				return diag.Errorf("error growing vkcs_db_cluster %s: %s", d.Id(), err)
-			}
-			stateConf.Pending = []string{string(dbClusterStatusShrink)}
-			stateConf.Target = []string{string(dbClusterStatusActive)}
-
-			_, err = stateConf.WaitForStateContext(ctx)
-			if err != nil {
-				return diag.Errorf("error waiting for vkcs_db_cluster %s to become ready: %s", d.Id(), err)
-			}
+		if sizeChange := new.(int) - old.(int); sizeChange > 0 {
+			err = databaseClusterActionGrow(updateCtx, "")
+		} else if sizeChange < 0 {
+			err = databaseClusterActionShrink(updateCtx, "")
+		}
+		if err != nil {
+			return databaseClusterUpdateProcessError(err, clusterID)
 		}
 	}
 
@@ -938,7 +747,7 @@ func resourceDatabaseClusterUpdate(ctx context.Context, d *schema.ResourceData, 
 			return diag.Errorf("unable to determine vkcs_db_cluster backup_schedule")
 		}
 
-		err = dbClusterUpdateBackupSchedule(DatabaseV1Client, d.Id(), &backupScheduleUpdateOpts).ExtractErr()
+		err = dbClusterUpdateBackupSchedule(dbClient, clusterID, &backupScheduleUpdateOpts).ExtractErr()
 
 		if err != nil {
 			return diag.Errorf("error updating backup schedule for vkcs_db_cluster %s: %s", d.Id(), err)
@@ -954,15 +763,10 @@ func resourceDatabaseClusterUpdate(ctx context.Context, d *schema.ResourceData, 
 	}
 
 	if d.HasChange("cloud_monitoring_enabled") {
-		_, new := d.GetChange("cloud_monitoring_enabled")
-		var cloudMonitoringOpts updateCloudMonitoringOpts
-		cloudMonitoringOpts.CloudMonitoring.Enable = new.(bool)
-
-		err := dbClusterAction(DatabaseV1Client, d.Id(), &cloudMonitoringOpts).ExtractErr()
+		err = databaseClusterUpdateCloudMonitoring(updateCtx)
 		if err != nil {
-			return diag.FromErr(err)
+			return databaseClusterUpdateProcessError(err, clusterID)
 		}
-		log.Printf("Update cloud_monitoring_enabled in vkcs_db_cluster %s", d.Id())
 	}
 
 	return resourceDatabaseClusterRead(ctx, d, meta)
@@ -995,4 +799,56 @@ func resourceDatabaseClusterDelete(ctx context.Context, d *schema.ResourceData, 
 	}
 
 	return nil
+}
+
+func databaseClusterUpdateProcessError(err error, clusterID string) diag.Diagnostics {
+	baseErr := err
+	if unwrappedErr := errors.Unwrap(err); unwrappedErr != nil {
+		baseErr = unwrappedErr
+	}
+
+	newErrMsg := baseErr.Error()
+	switch baseErr {
+	case errDBClusterNotFound:
+		newErrMsg = fmt.Sprintf("error retrieving vkcs_db_cluster %s", clusterID)
+	case errDBClusterUpdateWait:
+		newErrMsg = fmt.Sprintf("error waiting for vkcs_db_cluster %s to become ready", clusterID)
+
+	case errDBClusterUpdateDiskAutoexpand:
+		newErrMsg = fmt.Sprintf("error updating disk_autoexpand for vkcs_db_cluster %s", clusterID)
+	case errDBClusterUpdateDiskAutoexpandExtract:
+		newErrMsg = fmt.Sprintf("unable to determine disk_autoexpand from vkcs_db_cluster %s", clusterID)
+	case errDBClusterUpdateWalDiskAutoexpand:
+		newErrMsg = fmt.Sprintf("error updating wal_disk_autoexpand for vkcs_db_cluster %s", clusterID)
+	case errDBClusterUpdateWalDiskAutoexpandExtract:
+		newErrMsg = fmt.Sprintf("unable to determine wal_disk_autoexpand from vkcs_db_cluster %s", clusterID)
+	case errDBClusterUpdateCloudMonitoring:
+		newErrMsg = fmt.Sprintf("error updating cloud_monitoring_enabled for vkcs_db_cluster %s", clusterID)
+
+	case errDBClusterActionUpdateConfiguration:
+		newErrMsg = fmt.Sprintf("error updating configuration for vkcs_db_cluster %s", clusterID)
+	case errDBClusterActionApplyCapabitilies:
+		newErrMsg = fmt.Sprintf("error updating capabilities for vkcs_db_cluster %s", clusterID)
+	case errDBClusterActionApplyCapabilitiesExtract:
+		newErrMsg = fmt.Sprintf("error extracting capabilities for vkcs_db_cluster %s", clusterID)
+	case errDBClusterActionResizeWalVolumeExtract:
+		newErrMsg = fmt.Sprintf("unable to determine wal_volume from vkcs_db_cluster %s", clusterID)
+	case errDBClusterActionGrow:
+		newErrMsg = fmt.Sprintf("error growing vkcs_db_cluster %s", clusterID)
+	case errDBClusterActionShrink:
+		newErrMsg = fmt.Sprintf("error shrinking vkcs_db_cluster %s", clusterID)
+	case errDBClusterActionShrinkWrongOptions:
+		newErrMsg = fmt.Sprintf("invalid shrink options for vkcs_db_cluster %s", clusterID)
+	case errDBClusterActionShrinkInstancesExtract:
+		newErrMsg = fmt.Sprintf("error determining instances to shrink vkcs_db_cluster %s", clusterID)
+	case errDBClusterActionResizeVolume:
+		newErrMsg = fmt.Sprintf("error resizing volume for vkcs_db_cluster %s", clusterID)
+	case errDBClusterActionResizeWalVolume:
+		newErrMsg = fmt.Sprintf("error resizing wal_volume for vkcs_db_cluster %s", clusterID)
+	case errDBClusterActionResizeFlavor:
+		newErrMsg = fmt.Sprintf("error changing flavor for vkcs_db_cluster %s", clusterID)
+	}
+
+	errMsg := strings.Replace(err.Error(), baseErr.Error(), newErrMsg, 1)
+	return diag.Errorf(errMsg)
 }