+++ CHANGELOG.md
@@ -8,6 +8,7 @@ description: |-
 # VKCS Provider's changelog
 
 #### v0.2.1 (unreleased)
+- Support shards/instances/volumes resizing in DB cluster with shards
 - Fix reading state of DB cluster with shards
 
 #### v0.2.0